+++ src/cab.cpp
@@ -1,3 +1,3 @@
 #include "cab.h"
 
-Cab::Cab(int id, pair<int, int> location) : id(id), location(location) {}
+Cab::Cab(int id, pair<double, double> location) : id(id), location(location) {}
