+++ src/trip.cpp
@@ -1,3 +1,3 @@
 #include "trip.h"
 
-Trip::Trip(int id, pair<int, int> start_location) : id(id), start_location(start_location) {}
+Trip::Trip(int id, pair<double, double> start_location) : id(id), start_location(start_location) {}
